backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "exa-py>=1.12.1",
    "ipykernel>=6.29.5",
    "joblib>=1.4.2",
    "pip>=25.1",
    "python-dotenv>=1.1.0",
]
